change the queries to the new db's structure after the sprint I shall replace the queries with a new db interface

+++ alitheia/web-services/src/eu/sqooss/service/web/services/WebServices.java
@@ -159,7 +159,7 @@ public class WebServices {
      * @return the project or old one if exist
      */
     public WSStoredProject requestEvaluation4Project(String userName, String password,
-            String projectName, int projectVersion,
+            String projectName, long projectVersion,
             String srcRepositoryLocation, String mailingListLocation,
             String BTSLocation, String userEmailAddress,String website) {
         return webServices.requestEvaluation4Project(userName, password,
